refactor(admin): introduce RouteGroupNode component and enhance route management features
- Created a new RouteGroupNode component to encapsulate route group display and model ID copying functionality, improving code organization and reusability.
- Updated FlowBranch and FlowSection components to utilize the new RouteGroupNode, enhancing the overall structure of the route management interface.
- Adjusted layout and styling for better responsiveness and clarity in the RouteModelFlow and RouteSurfaceCatalog components.
- Enhanced translation keys to support new UI elements and improve user experience across multiple languages.

These changes aim to streamline route management and enhance usability in the admin interface.
